--- frontend/src/components/Navbar.css
@@ -1,118 +1,0 @@
-/* navbar.css */
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem;
-  background-color: rgb(242, 245, 41);
-}
-
-.navbar-left {
-  display: flex;
-  align-items: center;
-  position: relative;
-}
-
-.brand-logo {
-  font-size: 1.5rem;
-  font-weight: bold;
-  text-decoration: none;
-  color: black;
-}
-
-.dropdown-btn {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  height: 25px;
-  width: 35px;
-}
-
-.menu-icon {
-  width: 33px;
-  height: 30px;
-  background-color: inherit;
-  margin: 3px 0;
-  transition: transform 0.3s ease-in-out;
-}
-
-.dropdown-btn:focus {
-  outline: none;
-}
-
-.dropdown-content {
-  display: none;
-  flex-direction: column;
-  position: absolute;
-  top: 100%;
-  left: 0;
-  background-color: #f9f9f9;
-  min-width: 120px;
-  max-width: 250px;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-  z-index: 1;
-  font-size: 14px;
-  padding: 0;
-}
-
-.dropdown-content a {
-  text-decoration: none;
-  color: black;
-  padding: 8px 12px;
-  display: block;
-  white-space: nowrap;
-}
-
-.dropdown-content a:hover {
-  background-color: #e0e0e0;
-}
-
-@media screen and (max-width: 768px) {
-  .dropdown-content {
-    max-width: 180px;
-  }
-}
-
-.dropdown-content.open {
-  display: flex;
-}
-
-.navbar-right {
-  display: flex;
-  align-items: center;
-}
-
-.navbar-right > * {
-  margin-left: 20px;
-}
-
-.navbar-right > *:last-child {
-  margin-left: 0;
-}
-
-.logout-btn {
-  background-color: #e74c3c;
-  color: white;
-  border: none;
-  padding: 8px 16px;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.nav-link {
-  text-decoration: none;
-  color: black;
-  padding: 8px 12px;
-}
-
-.login-link {
-  background-color: inherit;
-  color: black;
-  border: none;
-  padding: 8px 16px;
-  border-radius: 4px;
-}
