+++ frontend/src/components/basic-ui/elements/Loader.js
@@ -1,0 +1,12 @@
+import classes from "./Loader.module.css";
+
+// Loader gif to show loading sign when a resource is being loaded
+function Loader() {
+  return (
+    <div className={classes.loader}>
+      <img src="/images/spinning-loading.gif" alt="Loading..." />
+    </div>
+  );
+}
+
+export default Loader;
