+++ frontend/src/components/basic-ui/elements/Loader.module.css
@@ -1,0 +1,13 @@
+.loader {
+  margin: 3rem auto;
+  width: 95%;
+  max-width: 100rem;
+  border-radius: 6px;
+  padding: 1rem;
+  text-align: center;
+}
+
+.loader img {
+  width: 300px;
+  height: 200px;
+}
