+++ frontend/src/components/basic-ui/elements/SkinColorPalette.css
