+++ frontend/src/components/basic-ui/elements/SkinColorPalette.js
