+++ frontend/src/components/basic-ui/elements/layout.module.css
@@ -1,0 +1,4 @@
+.main {
+    margin-top: 5rem;
+    padding-left: 3rem;
+}
