+++ frontend/src/components/user/FamilyMemberCard.css
