+++ frontend/src/components/user/LoginPage.js
@@ -1,0 +1,71 @@
+// LoginPage.js
+import React, { useState } from "react";
+import { Link } from 'react-router-dom'; // Import Link from react-router-dom
+import "./LoginPage.css"; // Import CSS file for login page styling
+
+function LoginPage() {
+  const [email, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState(null);
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      const response = await fetch("/login", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ email, password }),
+      });
+      if (!response.ok) {
+        throw new Error("Invalid credentials");
+      }
+      // Successful login - redirect to the home page or similar
+      window.location.href = "/";
+    } catch (error) {
+      setError(error.message);
+    }
+  };
+
+  return (
+    <div className="login-container">
+      <h2>Login</h2>
+      {error && <p className="error-message">{error}</p>}
+      <form onSubmit={handleSubmit} className="login-form">
+        <div className="form-row">
+          <label htmlFor="email" className="form-label">
+            Email
+          </label>
+          <input
+            type="email"
+            name="email"
+            id="email"
+            value={email}
+            onChange={(e) => setUsername(e.target.value)}
+            className="form-input"
+          />
+        </div>
+        <div className="form-row">
+          <label htmlFor="password" className="form-label">
+            Password
+          </label>
+          <input
+            type="password"
+            name="password"
+            id="password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            className="form-input"
+          />
+        </div>
+        
+        <button type="submit" className="login-btn">
+          Login
+        </button>
+      </form>
+      
+      <p>Not a user? <Link to="/register">Register here</Link></p>
+    </div>
+  );
+}
+
+export default LoginPage;
