+++ frontend/src/page404.js
@@ -1,0 +1,6 @@
+// Page 404 not found
+function Page404() {
+  return <div>Not found</div>;
+}
+
+export default Page404;
